Only call eips if fetch-dashboard succesfully completes
With this in place, fetch-dashboard can control whether or not eips will
be called. This is useful for implementing things like conditional
requests. In such case, fetch-dashboard could exit with a non-zero
status and prevent any calls to eips and prevent `num_refresh` to be
increased (which helps reduce the number of full screen refreshes).